+++ packages/bridge-ui/src/libs/bridge/Bridge.ts
@@ -1,10 +1,11 @@
 import { readContract, simulateContract, writeContract } from '@wagmi/core';
-import { Contract, utils } from 'ethers';
+import { Contract } from 'ethers';
 import { getContract, type Hash, UserRejectedRequestError, type WalletClient } from 'viem';
 
 import { bridgeAbi } from '$abi';
 import { routingContractsMap } from '$bridgeConfig';
 import { MessageStatusError, ProcessMessageError, ReleaseError, WrongChainError, WrongOwnerError } from '$libs/error';
+import { getEthersSigner } from '$libs/ethers/ethersClient';
 import type { BridgeProver } from '$libs/proof';
 import { getConnectedWallet } from '$libs/util/getConnectedWallet';
 import { getLogger } from '$libs/util/logger';
@@ -234,53 +235,44 @@ export abstract class Bridge {
     log('client', client);
 
     log('processing message', message, proof);
-    // const estimatedGas = await bridgeContract.estimateGas.processMessage([message, proof], { account: client.account });
-    // log('Estimated gas for processMessage', estimatedGas);
 
-    // // const wallet = await getConnectedWallet();
-
-    // // return await wallet.writeContract({
-    // //   address: bridgeContract.address,
-    // //   abi: bridgeContract.abi,
-    // //   functionName: 'processMessage',
-    // //   args: [message, proof],
-    // //   gas: 1000000n,
-    // // });
-
-    // const { request } = await simulateContract(config, {
-    //   address: bridgeContract.address,
-    //   abi: bridgeContract.abi,
-    //   functionName: 'processMessage',
-    //   args: [message, proof],
-    //   gas: estimatedGas,
-    // });
-    // log('Simulate contract for processMessage', request);
-
-    // return await writeContract(config, request);
-
-    // const estimatedGas = await bridgeContract.estimateGas.processMessage(message, proof, { from: client.account });
-    // console.log('Estimated gas for processMessage', estimatedGas.toString());
-
-    // Ensure proof is a valid bytes-like value
-    if (!proof || typeof proof !== 'string' || !proof.startsWith('0x')) {
-      throw new ProcessMessageError('Invalid proof format');
-    }
-
-    // Simplified proof handling (just for testing)
-    const simplifiedProof = utils.arrayify(proof);
-    log('Simplified Proof:', simplifiedProof);
-
-    const estimatedGas = 1_300_000;
-    log('Estimated gas for processMessage', estimatedGas.toString());
-
-    const signer = await getEthersSigner(config, { chainId: Number(bridgeTx.destChainId) });
-
-    // Instantiate the ethers contract using ABI and address from bridgeContract
-    const ethersBridgeContract = new Contract(bridgeContract.address, bridgeAbi, signer);
-    log('message', message);
-    log('message.data', message.data);
-
+    let estimatedGas;
     try {
+      estimatedGas = await bridgeContract.estimateGas.processMessage([message, proof], { account: client.account });
+      log('Estimated gas for processMessage', estimatedGas);
+
+      // const wallet = await getConnectedWallet();
+
+      // return await wallet.writeContract({
+      //   address: bridgeContract.address,
+      //   abi: bridgeContract.abi,
+      //   functionName: 'processMessage',
+      //   args: [message, proof],
+      //   gas: 1000000n,
+      // });
+
+      const { request } = await simulateContract(config, {
+        address: bridgeContract.address,
+        abi: bridgeContract.abi,
+        functionName: 'processMessage',
+        args: [message, proof],
+        gas: estimatedGas,
+      });
+      log('Simulate contract for processMessage', request);
+
+      return await writeContract(config, request);
+    } catch (e) {
+      // If we get an error, we fall back to ethers.js and try again
+      console.error(e);
+      log('Error in processNewMessage, falling back to ethers.js', e);
+
+      const signer = await getEthersSigner(config, { chainId: Number(bridgeTx.destChainId) });
+
+      const ethersBridgeContract = new Contract(bridgeContract.address, bridgeAbi, signer);
+
+      estimatedGas = ethersBridgeContract.estimateGas.processMessage(message, proof);
+      log('Estimated gas for processMessage', estimatedGas.toString());
+
       const tx = await ethersBridgeContract.processMessage(message, proof, {
         gasLimit: estimatedGas,
       });
@@ -288,9 +280,6 @@ export abstract class Bridge {
       const response = await tx.wait();
 
       return response.transactionHash;
-    } catch (error) {
-      console.error('Transaction error:', error);
-      throw new ProcessMessageError('Transaction failed');
     }
   }
 
@@ -343,52 +332,4 @@ export abstract class Bridge {
 
     return await writeContract(config, request);
   }
-}
-
-import { type Config, getClient } from '@wagmi/core';
-import { providers } from 'ethers';
-import type { Chain, Client, Transport } from 'viem';
-
-export function clientToProvider(client: Client<Transport, Chain>) {
-  const { chain, transport } = client;
-  const network = {
-    chainId: chain.id,
-    name: chain.name,
-    ensAddress: chain.contracts?.ensRegistry?.address,
-  };
-  if (transport.type === 'fallback')
-    return new providers.FallbackProvider(
-      (transport.transports as ReturnType<Transport>[]).map(
-        ({ value }) => new providers.JsonRpcProvider(value?.url, network),
-      ),
-    );
-  return new providers.JsonRpcProvider(transport.url, network);
-}
-
-/** Action to convert a viem Public Client to an ethers.js Provider. */
-export function getEthersProvider(config: Config, { chainId }: { chainId?: number } = {}) {
-  const client = getClient(config, { chainId });
-  if (!client) return;
-  return clientToProvider(client);
-}
-
-import { getConnectorClient } from '@wagmi/core';
-import type { Account } from 'viem';
-
-export function clientToSigner(client: Client<Transport, Chain, Account>) {
-  const { account, chain, transport } = client;
-  const network = {
-    chainId: chain.id,
-    name: chain.name,
-    ensAddress: chain.contracts?.ensRegistry?.address,
-  };
-  const provider = new providers.Web3Provider(transport, network);
-  const signer = provider.getSigner(account.address);
-  return signer;
-}
-
-/** Action to convert a Viem Client to an ethers.js Signer. */
-export async function getEthersSigner(config: Config, { chainId }: { chainId?: number } = {}) {
-  const client = await getConnectorClient(config, { chainId });
-  return clientToSigner(client);
 }
